stock permanen don

+++ app/Http/Controllers/ProductController.php
@@ -80,6 +80,10 @@ class ProductController extends Controller
     {
         $product = Product::findOrFail($id);
         
+        if ($product->is_permanent) {
+            return redirect()->back()->with('error', 'This product cannot be deleted!');
+        }
+
         // Delete image if exists
         if ($product->image && $product->image !== 'images/contoh2.png') {
             Storage::disk('public')->delete($product->image);
